% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIMseq.publication-package.R
\docType{package}
\name{CIMseq.publication-package}
\alias{CIMseq.publication-package}
\alias{CIMseq.publication}
\title{Reproduction of the CIMseq publication figures.}
\description{
\tabular{ll}{ Package: \tab CIMseq.publication\cr Type: \tab Package\cr
Version: \tab 1.0\cr Date: \tab 2016-02-28\cr License: \tab GPL-3\cr }
}
\author{
Author: Jason T. Serviss
}
\keyword{package}
