% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getCountsDataViaLinks}
\alias{getCountsDataViaLinks}
\title{getCountsDataViaLinks}
\arguments{
\item{links}{character; The links to the counts files to download.}
}
\value{
A data.frame with the unfiltered counts data.
}
\description{
Downloads the count data project and concatenates it into a data.frame. 
The column containing the gene name, to be named HGN, is expected in all 
counts files.
}
\author{
Jason Serviss
}
