% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getMetadataViaLink}
\alias{getMetadataViaLink}
\title{getMetadataViaLink}
\arguments{
\item{link}{character; Link to the processed metadata.}
}
\value{
The metadata tibble containing the concatenated information in the 
project annotation folder.
}
\description{
Utilizes the EngeMetadata package to download and format the metadata
associated with a specific project folder. Downloads all metadata/plate info
present in the annotation folder by default.
}
\author{
Jason Serviss
}
